From minima to immaculate?
Attempting to change the theme -- it's a fail; I just learned this is not something you can yet do in jekyll, but something later implementations will make possible. At the moment, your best bet is to pick a theme, download it, and copy your material into its structure. Seems silly.